Block 12: split 4 oversized Rust command modules
- commands/skills.rs (759) → skills/{mod,search,install}.rs
- commands/pty.rs (686) → pty/{mod,spawn,stream}.rs
- commands/health.rs (630) → health/{mod,deps,run}.rs
- commands/projects/mod.rs (566) → kept cross-cutting, extracted
  ui_state.rs + dev_server.rs, renamed windows.rs → window_registry.rs

No public API changes. No #[tauri::command] names changed so lib.rs
invoke_handler and all frontend invoke() calls still resolve. 119 Rust
tests pass first try; typecheck clean.
